+++ bind/pkg/client/clientset/versioned/typed/bind/v1alpha1/generated_expansion.go
@@ -26,4 +26,6 @@ type APIServiceExportRequestExpansion interface{}
 
 type APIServiceNamespaceExpansion interface{}
 
+type AppClusterBindingExpansion interface{}
+
 type ClusterBindingExpansion interface{}